requirements.txt
astropy==6.1.3
joblib==1.4.2
keras_tuner==1.4.7
numpy==2.3.2
openai==1.99.6
pandas==2.3.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.1
scipy==1.16.1
tensorflow==2.16.2
tf==1.0.3
tqdm==4.66.5
